backend/requirements.txt
soundcard>=0.4.0
websockets>=14.0
numpy>=1.24.0
winrt-Windows.Media.Control>=3.2.0
requests>=2.28.0
Pillow>=9.0.0
pyacoustid>=1.3.0
